requirements.txt
flask
flask-cors
mysql-connector-python
python-dotenv
python-decouple
gunicorn
google-auth
requests
PyJWT==2.9.0
Pillow>=10.0.0
qrcode[pil]
geopy
pandas
openpyxl>=3.1.2
flask-limiter
werkzeug
pydantic>=2.0.0
